--- src/Config/Scenes.js
@@ -1,22 +1,0 @@
-import Preload from "../Scenes/Preload";
-import Home from "../Scenes/Home";
-import Start from "../Scenes/Start";
-import Options from "../Scenes/Options";
-import Credits from "../Scenes/Credits";
-import Loading from "../Scenes/Loading";
-import Pascal from "../Scenes/GamePlay/Pascal";
-import VisualBasic from "../Scenes/GamePlay/VisualBasic";
-
-const Scenes = [
-	Preload,
-	Home,
-	Start,
-	Options,
-	Credits,
-	Loading,
-	// Game Play
-	Pascal,
-	VisualBasic,
-];
-
-export default Scenes;
