+++ src/Scenes/GamePlay/VisualBasic.js
@@ -1,4 +1,4 @@
-import GlobalConfigs from "../../Config/Configs";
+import GlobalConfigs from "../../Configs";
 import { randomNumber } from "../../Utils/Maths";
 
 import Player from "../../Objects/ShootUp/Player";
