+++ src/Scenes/Home.js
@@ -1,50 +1,84 @@
-import GlobalConfigs from "../Config/Configs";
+import GlobalConfigs from "../Configs";
 
-import { TextStyle } from "../Theme";
+import AnimateTitle from "../Components/AnimatedTitle";
+
+import Player from "../Objects/Platform/Player";
+import Tiles from "../Objects/Platform/Tiles";
+import DirectionSign from "../Objects/DirectionSign";
 
 export default class Home extends Phaser.Scene {
 	constructor() {
 		super({ key: "Home" });
 	}
 
+	init(data) {
+		this.lastScene = data.sceneName;
+	}
+
 	create() {
+		this.createWorld();
+		this.createPlayer();
+		this.createCollision();
+	}
+
+	createWorld() {
 		const { width, height, middleWidth, middleHeight } = GlobalConfigs.screen;
 
-		const margin = 40;
+		// -- Background
+		const background = this.add.image(middleWidth, middleHeight, "BackgroundForest").setScale(1.3, 1);
 
-		// Start
-		const configStart = {
-			x: middleWidth,
-			y: middleHeight - margin,
-			text: "Start",
-			action: () => this.scene.start("Start"),
-		};
-		this.createText(configStart);
+		{	// -- Floor
+			this.tilesGroup = this.physics.add.staticGroup({ classType: Tiles, });
+			const grassNumber = width / 128;
+			for (let i = 0; i < grassNumber; i++) {
+				const g = this.tilesGroup.get(128 * i + 64, height, 1);
+			}
+		}
 
-		// Old
-		const configOldSite = {
-			x: middleWidth,
-			y: middleHeight + margin,
-			text: "Old Site",
-			action: () => this.openOldSite(),
-		};
-		this.createText(configOldSite);
+		{	// -- Objects
+			const x = 50;
+			const y = height - 90;
+
+			const signDirectionStart = new DirectionSign(this, width - x, y, "Game");
+			signDirectionStart.changeDepth(1);
+			const signDirectionOptions = new DirectionSign(this, x, y, "Options").setFlipX(true);
+			signDirectionOptions.changeDepth(1);
+			signDirectionOptions.label.x += 10;
+
+			const tree1 = this.add.image(middleWidth, height - 86, "Tree1");
+			const tree2 = this.add.image(middleWidth / 2, height - 200, "Tree2").setDepth(20);
+			const tree3 = this.add.image(width * 0.75, height - 200, "Tree2").setDepth(20);
+		}
+
+		{ // UI
+			const title = new AnimateTitle(this, middleWidth, 100, "201flaviosilva");
+		}
 	}
 
-	openOldSite() {
-		const url = "https://201flaviosilva.bitbucket.io/src/201flaviosilva/V1/index.html";
-		const s = window.open(url, '_blank');
+	createPlayer() {
+		const { width, height, middleWidth, middleHeight } = GlobalConfigs.screen;
+		this.playersGroup = this.physics.add.group({
+			classType: Player,
+			collideWorldBounds: true,
+			runChildUpdate: true
+		});
 
-		if (s && s.focus) s.focus();
-		else if (!s) window.location.href = url;
+		const x = this.lastScene === "Options" ? 50 : middleWidth;
+		this.player = this.playersGroup.get(x, middleHeight);
+		this.player.generate();
 	}
 
-	createText(configs) {
-		const { x, y, text, action } = configs;
-		const label = this.add.text(x, y, text, TextStyle.home.menu.normal).setOrigin(0.5);
-		label.setInteractive({ useHandCursor: true });
-		label.on('pointerover', () => label.setStyle(TextStyle.home.menu.over));
-		label.on('pointerout', () => label.setStyle(TextStyle.home.menu.normal));
-		label.on('pointerup', action);
+	createCollision() {
+		this.physics.add.collider(this.tilesGroup, this.player);
+	}
+
+	update() {
+		if (this.player.x < this.player.width * 0.75) {
+			this.scene.start("Options", { sceneName: "Start" });
+			this.scene.stop();
+		} else if (this.player.x > GlobalConfigs.screen.width - this.player.width * 0.75) {
+			this.scene.start("Loading", { nextGame: "Pascal" });
+			this.scene.stop();
+		}
 	}
 }
